requirements.txt
aiohttp==3.9.4
ffmpy==0.3.1
health_ping==1.0.1
i18nice==0.6.2
Pillow==11.2.1
pyheif==0.7.1
python-dotenv==1.0.1
python-telegram-bot[job-queue]==22.1
rlottie_python==1.2.1
